[Workflow 적용을 위한 프로젝트 구조 재구성]
- youtube 수집 모듈 함수 분리
- portal 수집 모듈 함수 분리
- config 경로 정리
- news 수집 모듈 naver 검색 api 활용하도록 변경
- blog/news 모듈 중복 함수 통합
- youtube 모듈 추가(좋아요 수, 조회 수, 제목, 썸네일 수집)
- git ignore에 tmp_files 경로 추가
